Ability to connect to Redis cluster using SSL/TLS
Added `REDIS_SCHEME` setting to be able to specify whether connection
with Redis server should be secure (`rediss`) or nor (`redis`).

This defaults to `redis` for retrocompatibility but can be set to
`rediss` (notice the additional `s` as in "secure") to allow use
encryption in-transit when this is enabled on the Redis cluster.

This is necessary necessary to connect to the Redis cluster on AWS
ElastiCache which uses TLS for encryption in transit.

Also set `timeout=30` in Django Channels' settings as by default never
times out (which is never a great idea, especially on production).

Part of ticket: https://trello.com/c/YG9anK8A