CI: Replacing CircleCI with GitHub Actions
Why?
====

We're aiming to simplify our current CI/CD and reduce the number of
technologies involved.
Ideally we'd like to use GitHub Actions for all things CI/CD.
This is a step in that direction and remove CircleCI out of the equation.

What
====

The workflow is very very similar to old CircleCI workflow with
few tweaks:
- Docker caching is done using the `satackey/action-docker-layer-caching`
  action which is doing the same docker save/load under the hood
- `CIRCLE_SHA1`/`CIRCLE_BRANCH` replaced with more long winded but
  equivalent way of getting the same information in GH Actions
- Quay.io username/password stored in GH Repository secrets instead of
  being stored in CircleCI
- Speaking of secrets, the credentials/robot for this GH Actions are new
  and different from old CircleCI one to avoid any possible confusion

NOTE: `npm install` in `testjs` job is cached but the
benefits are very small as non-cached `npm-instal` was
taking 16-20s anyway.

Further splitting in smaller jobs
=================================

I tried to split the current final job into smaller, independent, jobs which
were:
- building the docker image
- running the tests with docker-compose
- push to quay (once tests passed)

Bear in mind different jobs don't share state.

I tried to use GH Actions artifacts to store the docker image but this
was slow because of size of docker image and upload/download required
to transfer this between jobs.
In the end the pragmatic thing to do was to do all these things in the
final job together (which was also the approach taken in the old
CircleCI workflow).

On branch name and commit hash
==============================

Unfortunately GH Actions doesn't seem as polished as CircleCI
when it comes to getting basic info like branch name and commit hash.

There is no easy way (that I found) to get these info by using a simple
environment variable and even the available variables like `GITHUB_REF`,
`GITHUB_HEAD_REF` and `GITHUB_SHA` have different values depending on
what triggered the build - how confusing!!

This is the reason why the `Get branch name (when xxx)` and
`Get commit SHA1 (when xxx)` are not as simple as I'd like.

These steps use may run on not depending on what triggered the build,
e.g. a Pull Request or else (hence the `if: github.event_name ...`)

To make things worse sometimes to get the correct info you have to get
it from the event payload like in the case of the commit hash when it's
a PR (`github.event.pull_request.head.sha`) or otherwise you'll get an
odd commit hash instead.

See resources I've found to help me get these info for more details:
- Get branch name: https://stackoverflow.com/a/64210623/455642
- Get correct commit hash: https://github.community/t/github-sha-isnt-the-value-expected/17903

Ticket
======

Ticket: https://trello.com/c/jyRlofox