Configure dependabot to ignore node dependencies
After enabling dependabot, it insists on checking node dependencies for what are
development only tools.  This will result in a lot of churn for issues that are
highly unlikely to affect us in production.  We are more interested in just the
python dependencies and code that will run in production.

Adding the dependabot config should enable us to be specific about what we want
to check, without it making assumptions because of the presence of the node
package.json file.