This adds the prometheus metrics view.
Its not the recommended pattern from the library documentation which involved appending their urls.py to the base of the url patterns, which felt like a recipe for disaster because there was nothing to stop a bad commit from breaking the rest of the platform, so I took the individual line out of their URLs file.

Also the task specified that it should be at `/metrics`, and their solution would put it at `/prometheus/metrics`.