Enables the Django admin
Currently, the control panel does not use the Django Admin app for
reasons lost in the mists of time.  Whilst the admin app would be wrong
if used as object scaffolding, for administering the running system, it
seems perfect.  In cases where we want to make changes to objects, in
bulk, it makes more sense to allow our PM/DM to do the work, than
requiring a developer.

In addition to enabling the django admin, this PR will also add an
action to user objects, allowing all of the selected users to be marked
ready for migration.  Users to be reset to the old system will need to
be edited directly.