Update the text on the login fail page
The login fail page shows the user a lot of reasons that they could have
ended up on that page, but the most likely reason (timeout) is buried in
a paragraph of text.

This PR moves the most likely reason to the top, to make it easier to
find the login link.