Simplify the on_authenticated process in EKS
When users log in to the EKS control panel, the `on_authenticate` method
is executed to determine the current migration state of the user and run
the appropriate helm charts for new or migrated users.

Previously the many states in this method were confusing and obscuring
some edge cases, and so this PR attempts to simplify the method by
following the 'Return Early' pattern of getting out of the method as
soon as possible.  This results in code that is hopefully easier to read
and update when we only have the EKS environment.

Whereas the previous code executed a `helm.delete` step on the previous
init charts, it did it within this namespace.  Presumably if a user is
migrating, their init charts are on alpha, rather than EKS?  It isn't
yet clear whether that step was necessary, and if so how we should
handle failure.