Stops delete app from using helm in wrong cluster
Currently in the EKS cluster, attempting to delete an App (model)
results in an attempt to use helm to delete the app. The breakage
happens because the app was deployed by concourse in a different
cluster which is inaccessible to this one.

By removing call to ask Helm to delete the app, we are abdicating
responsibility for the cleanup of the K8s resources, which must then be
handled another way.