Adds a new command for adding email address to apps
Adds a new command that will, when given an email address and an app
slug, ensure that a user with the email has access to that app.

In the above context, a user is an auth0 user who is registered under an
email address and there the 'connection' is 'email' (as opposed to
github for most AP users).  An app is a 'group' in the old Auth0
Authorization extension API.

The controlpanel.api.auth0.AuthorizationAPI class is used to interact
with the auth0 extension which is rather underdeveloped. To add a user
id to an app it will first iterate through every group in an attempt to
find one with a matching name before looking up the email and then
PATCHing the group members.  This is rather inefficient so we will do
this in separate stages.