Adds a simple authenticated page for development
During development it is difficult to try out new features, or changes
to existing features without building something that will be
accidentally shipped to production. As we do not currently have
effective feature flags that can be controlled at runtime, this PR
introduces a tactical solution to enable us to make progress.

A new django app `develop` is introduced which is only installed, and
whose routes are only connected, when the system is run using the
`controlpanel.settings.development` configuration file.

Currently the new app only contains a single view, `develop_index` which
in development is available at http://127.0.0.1:8000/develop/. Whilst it
is expected this may in future be used to link to other trial features,
for now it provides an interface that mocks up a rather simple tools
page allowing for:

* Selection by the user of a single tool and submission back to the
  endpoint
* Listing of currently installed tools for the user

The page requires authentication, and therefore during development the
user will be required to 2fa via github, making the current username
available in `request.user`.