Remove settings.EKS flag
During the 2021/2022 KOPS->EKS migration there were some subtle changes
in behaviour and these differences were captured in if/else blocks based
on a configuration setting (`EKS` commonly used with `if settings.TKS`).
Whilst this could possibly have been modelled a little differently it
fulfilled its purpose in abstracting away the differences between the
two types of clusters.

This change removes all of the if/else blocks predicated on the value of
settings.EKS leaving in place just the EKS based code. Tests have also
been changed to remove the patching of the flag and the now duplicated
tests (for KOPs and EKS).