Install sass to replace node-sass package
Around December 2020, libsass (node-sass) was discontinued and
was replaced with Dart Sass (sass).

I have been unable to install the npm packages due to this version being in use,
it would simply not compile.

Dependabot tried to do its best on PR #992, by upgrading the package to version 7.0.0,
although that seemed to have broken the test suite and was never looked into since.

This is an attempt to make the replacement, it probably needs accomodation
in case the sass API has changed.

If the test suite passes at the Pull Request level, please do test on a local environment.