Simplify the register app form (#1228)
* Simplify the register app form

Only ask for a repo URL and datasource when registering an app.
This is because the form will now be publically accessible,
with a new flow for creating auth0 clients. Also removes JS to
lookup repositories, as this is no longer needed.

App registration form should be open to logged in users, to help
enable self service of app registration. Updating connections
remains superuser only.

* Update app assume role policy

Adds statement to allow cloud platform to assume the app IAM role

* Check repo url on app registration API endpoint