Bugfix/namespace length (#1257)
* Add namespace field to App

Will be used to store the Cloud Platform namespace name. Adds migrations
that add the field, then a data migration to add values for existing apps,
then finally a third migration to require values when creating a new
object.

* Add namespace field to the register app form

Adds the field in the HTML.
Update the OIDC template to use the entered namespace when
creating the app IAM role.

* Add APP_ROLE_ARN secret after creating role

Updates the create app role task handler to pass a github api
token, so this can be used to create the APP_ROLE_ARN secret as
soon as the App role is created. This is required for apps that
will be deployed without Auth0 clients.