Feature/deprecate tool (#1404)
* Add is_deprecated flag to Tool model

* Add field to mark release retired and add deprecation message

* Disable open tool button

* Handle retired and deprecated tools in the frontend

- If a tool is depcrecated, show the deprecation message when it is selected
- If a tool is retired, hide it and show a message that user must upgrade

* Display deprecation message as a warning

* Show retired message as a warning

* Disable buttons based on selected release

If a deployed release is selected, enable open and restart button.
If a undeployed release is selected, enable deploy button only.

* Replace image_tag property with model field

* Add image_tag to the release detail, create pages

Use the value stored in the DB when deploying the tool

* Make tool description a required field

* Fix queryset when looking form related tool

Previously all restricted tools were excluded. However, allowing tools
to be deprecated means we can update this logic. As some restricted tools
may not be deprecated, so should not display an "unsupported" message.

* Fix bug finding tools using rc chart version

* Filter tool releases by status in superuser view

* Add status tag in release admin, and allow filtering by status