Convert ToolDeployment to a django model
Allows us to track the specific tool release that
a user has deployed.
This commit includes minimal changes to get the
model working, so that every time a tool is
deployed from the front end, a database object is
created. This does introduce a bug about status of
the tool deployment, but this will be fixed later.