Track if a deployment is active
Use an is_active field to allow us to track past
deployments.
This also makes it easier to uninstall an old tool
before we install a new one.

NB a bug will occur if a user has no active
ToolDeployment but tries to switch from a jupyter
datascience to all-spark release (or vice versa).
To resolve, we have to manually install the users
Jupyter release, or deploy the same type of
Jupyter to create an "active" ToolDeployment
object, and then the user can deploy the other
type of Jupyter.