Update create user to expect users from Entra (#1429)
* Update create user to expect users from Entra

To enable CICA users to access the CP, we will
enable a new connection in Auth0 with EntraID.
This updates the code we use to create a user from
the ID token to normalize their name, and store
their email as the justice_email, to avoid them
having to reauth to capture it.

* Use users slug rather than username in helm

Previously the username would be used when installing user helm charts.
This was fine when all users came from github, as their usernames were
guaranteed to be valid with helm. However this is not the case with
usernames from Entra, as they can include invalid characters such as '.'
which results in an error after the user logs in trying to provision
the user. This changes uses the slug, which is valid for helm, however
accessing tools will still not be compatible so tooling will not be
available for these users.

* Improve justice email validation

* Revert "Use users slug rather than username in helm"

This reverts commit 93cb6cdabc70890504042d7eab7060b2ce2b81ff.

We will resolve the issue these changes were meant to fix in a different
way when we fully implement access with EntraID

* Limit access unless user is created via GitHub auth

* Additional test for creating user

* Fix typo

* Fix failing tests