Update script to create tool deployments (#1448)
Previously it ignored tools that had been retired.
A dry run in prod showed this would be a
significant number. It would be better to still
have a record of these tool deployments, so we
can see how many users are using retired releases
and potentially reach out to them.