Make sure helm install and delete set returncode
Previously we would always read from stderr and stdout when running a
helm install/delete command. This would indirectly cause the subprocess
to 'wait', because it would wait until the EOF was reached. But this
wasnt clear, so previous changes broke deployments as I was relying on
a returncode being present but it no longer was as the calls to stderr
happened later.

So now we explicitly have the subprocess wait for a returncode. There is
still a bug present n how the timeout is passed, as it is not used by
helm, only the subprocess. But for now it shouldnt cause issues as helm
completes before the subprocess timeout is reached.