Refactor tasks to delete tools
Delay the tool uninstall until 3am the next day, to avoid uninstalling
a release when people are using them. Delay is only enabled in prod, in
dev they are uninstalled immediately.

Remove dependency on an instance in the task that calls the helm
uninstall. This allows us to uninstall tools after they have been
deleted from the database.