Feature/uninstall retired releases (#1462)
* Use celery to uninstall retired tools

Delay the tool uninstall until 3am the next day, to avoid uninstalling
a release when people are using them. Delay is only enabled in prod, in
dev they are uninstalled immediately.

Remove dependency on an instance in the task that calls the helm
uninstall. This allows us to uninstall tools after they have been
deleted from the database.

* Refactor how helm executes commands

* Make sure helm install and delete set returncode

Previously we would always read from stderr and stdout when running a
helm install/delete command. This would indirectly cause the subprocess
to 'wait', because it would wait until the EOF was reached. But this
wasnt clear, so previous changes broke deployments as I was relying on
a returncode being present but it no longer was as the calls to stderr
happened later.

So now we explicitly have the subprocess wait for a returncode. There is
still a bug present n how the timeout is passed, as it is not used by
helm, only the subprocess. But for now it shouldnt cause issues as helm
completes before the subprocess timeout is reached.

* Remove codeql