Update logging errors to warnings
Sentry will create an issue for any log.error call. There are some cases
where we log an error that is not error worthy - for example, where we
try to delete a release that is not found. We handle this in the code,
so update this output to a warning rather than an error to keep Sentry
less noisy.