Update logging errors to warnings (#1481)
* Update logging errors to warnings

Sentry will create an issue for any log.error log.warning call. There are some cases
where we log an error that is not error worthy - for example, where we
try to delete a release that is not found. We handle this in the code,
so update this output to info to keep Sentry less noisy.