Bugfix: Make admin field required in dashboard detail page (#1518)
Make admin field required in dashboard detail page

Add JS validation to sure a choice is made in the admin select field.
To enable this I had to amend the autocomplete.js to check for a
required data attribute.
Also add server side validation to ensure that an admin is selected,
otherwise redirect with an error message.