Further improvements to handling Helm errors (#1650)
This commit makes two changes to how we handle Helm errors when
deploying tools:

1. Fix a bug introduced in previous commit dealing with transient errors
2. Catch timeout errors and raise them specifically

For the first change, by returning None after handling transient errors,
we avoid a bug where the code downstream tries to read outputs from
the subprocess. This error can be seen in Sentry:
https://ministryofjustice.sentry.io/issues/6764901715/?environment=dev&project=5606313&query=&referrer=issue-stream

The second change is intended to seperate timeout errors from other Helm
errors, to improve visibility of errors in Sentry.

Together these changes should reduce the number of errors that are
raised in Sentry, and make those that are raised easier to interpret.