Bugfix/helm in progress error (#1654)
* Add exception for Helm operation in progress error

This is a known exeption, and is generally transient, but when it gets
raised as a standard ToolDeploymentError it was creating unnessary noise
and made it harder to spot real issues. So by raising a specific exception
we can supress it in the Sentry UI, but still keep a record of it happening
so that if the number of alerts spikes we can investigate further.

* Remove unused code